* Added: Direct3D 8 renderer

+++ src/RenderList.h
@@ -32,7 +32,7 @@ struct Sprite
 
 struct Color
 {
-    f32 iR, iG, iB, iA;
+    u32 iR, iG, iB, iA;
 };
 
 struct UpdateData
